% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inla_mats.R
\name{create_inla_mats}
\alias{create_inla_mats}
\title{Create the matrices required for SPDE approximation in CensSpBayes}
\usage{
create_inla_mats(S,S_pred=NULL,
                         offset=NULL,
                         cutoff=1e-12,
                         max.edge=NULL,...)
}
\arguments{
\item{S}{Matrix of locations where data is observed}

\item{S_pred}{(optional) Matrix of locations where data is to be predicted. If this is not supplied, the corresponding index matrix will be returned as a NULL object}

\item{offset}{The automatic extension distance. One or two values, for an inner and an optional outer extension. If negative, interpreted as a factor relative to the approximate data diameter (default=-0.10???). Same as in fmesher::fm_mesh_2d_inla()}

\item{cutoff}{The minimum allowed distance between points. Point at most as far apart as this are replaced by a single vertex prior to the mesh refinement step. Same as in fmesher::fm_mesh_2d_inla()}

\item{max.edge}{The largest allowed triangle edge length. One or two values. Same as in fmesher::fm_mesh_2d_inla()}

\item{...}{Additional parameters to be passed to the function fmesher::fm_mesh_2d_inla()}
}
\value{
list of matrices needed for the SPDE approximation. The entries are: c.mat(mass matrix), g1.mat(stiffness matrix), g2.mat(a combination of mass and stiffness matrices), A(index matrix for observation locations), A.pred(index matrix for prediction locations)
}
\description{
Create the mass and stiffness matrices and the index matrices for both the observation locations and the prediction locations
}
