% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{simdat: A simulated spatial data with left-censoring}
\format{
\subsection{\code{simdat}}{

A data frame with 10,000 rows and 4 columns:
\describe{
\item{Y}{The response, censored whenever this is smaller than cutoff.Y}
\item{cutoff.Y}{The limit below which any data is censored}
\item{x, y}{x and y locations of the observations}
}
}
}
\usage{
simdat
}
\description{
Simulated Dataset to use the CensSpBayes function on
}
\keyword{datasets}
